Add db_encrypted_password parameter
Unfortunately, newer versions of Artifactory will modify db.properties
on startup if it finds a non artifactory-encrypted database password.
There doesn't appear to be a way to supress this behaviour.  When puppet
next runs, it sees the config file has changed, puts it back and
restarts artifactory.  Artifactory then reencrypts the value and this
repeats every puppet run.

This commit adds an optional new `db_encrypted_password` parameter.  If
this is set, it will be used instead of `db_password`.  This is far from
perfect, but I couldn't think of a sane way of automatically calculating
`db_encrypted_password` from `db_password`.  I could probably work out
how the encryption is done *and* write a puppet function, but we'd
either need to also add a `master_key` parameter (and overwriting the
artifactory generated one is probably a *bad* idea as its used to
encrypt other passwords), or collect the current `master_key` as a fact
(which probably isn't a great idea either).

Fixes #10